# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: M International AI Quote Agent operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /emails/{emailId}:
    get:
      operationId: getEmail
      tags: [ingestion]
      summary: Get an email and its processing details
      parameters:
        - $ref: "#/components/parameters/EmailId"
      responses:
        "200":
          description: Email detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailDetail"
        "404":
          description: Email not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /emails/{emailId}/process:
    post:
      operationId: processEmail
      tags: [ingestion]
      summary: Process an email through the RFQ analysis pipeline
      parameters:
        - $ref: "#/components/parameters/EmailId"
      responses:
        "200":
          description: Processing result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailProcessResult"
        "404":
          description: Email not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /email-attachments/{attachmentId}:
    get:
      operationId: getEmailAttachment
      tags: [ingestion]
      summary: Get attachment metadata
      parameters:
        - $ref: "#/components/parameters/AttachmentId"
      responses:
        "200":
          description: Attachment metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailAttachment"
        "404":
          description: Attachment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /ai-review/{rfqId}:
    get:
      operationId: getAiReview
      tags: [ingestion]
      summary: Get AI review detail
      parameters:
        - $ref: "#/components/parameters/RfqId"
      responses:
        "200":
          description: AI review detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiReviewDetail"
        "404":
          description: Review item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ai-review/{rfqId}/approve:
    post:
      operationId: approveAiReview
      tags: [ingestion]
      summary: Approve an AI result
      parameters:
        - $ref: "#/components/parameters/RfqId"
      responses:
        "200":
          description: Updated AI review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiReviewDetail"
  /ai-review/{rfqId}/reclassify:
    post:
      operationId: reclassifyAiReview
      tags: [ingestion]
      summary: Change the request classification
      parameters:
        - $ref: "#/components/parameters/RfqId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReclassifyReviewInput"
      responses:
        "200":
          description: Updated AI review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiReviewDetail"
  /ai-review/{rfqId}/reject:
    post:
      operationId: rejectAiReview
      tags: [ingestion]
      summary: Reject an AI result as non-RFQ
      parameters:
        - $ref: "#/components/parameters/RfqId"
      responses:
        "200":
          description: Updated AI review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiReviewDetail"
  /ai-review/{rfqId}/update:
    post:
      operationId: updateAiReview
      tags: [ingestion]
      summary: Edit extracted RFQ details
      parameters:
        - $ref: "#/components/parameters/RfqId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReviewInput"
      responses:
        "200":
          description: Updated AI review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiReviewDetail"
...
components:
  parameters:
    RfqId:
      name: rfqId
      in: path
      required: true
      schema:
        type: integer
    EmailId:
      name: emailId
      in: path
      required: true
      schema:
        type: integer
    AttachmentId:
      name: attachmentId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    RfqSource:
      type: string
      enum: [EMAIL, WEBSITE, PHONE, PORTAL]
    RequestType:
      type: string
      enum: [PARTS_EXCHANGE, NEW_PART_PURCHASE, REPAIR, OVERHAUL, INSPECTION, UNKNOWN]
    RfqStatus:
      type: string
      enum: [NEW, ANALYZING, UNDER_REVIEW, PROCESSING, NEEDS_HUMAN_REVIEW, COMPLETED, CLOSED]
    Priority:
      type: string
      enum: [LOW, MEDIUM, HIGH, URGENT]
...
    Rfq:
      type: object
      properties:
        id:
          type: integer
        rfqNumber:
          type: string
        customer:
          type: string
        customerCompany:
          type: string
        partNumber:
          type: string
        source:
          $ref: "#/components/schemas/RfqSource"
        requestType:
          $ref: "#/components/schemas/RequestType"
        confidence:
          type: integer
        priority:
          $ref: "#/components/schemas/Priority"
        status:
          $ref: "#/components/schemas/RfqStatus"
        createdAt:
          type: string
        age:
          type: string
      required: [id, rfqNumber, customer, customerCompany, partNumber, source, requestType, confidence, priority, status, createdAt, age]
...
    Email:
      type: object
      properties:
        id:
          type: integer
        sender:
          type: string
        senderEmail:
          type: string
        subject:
          type: string
        receivedAt:
          type: string
        aiStatus:
          type: string
          enum: [ANALYZED, PROCESSING, FAILED]
        requestType:
          $ref: "#/components/schemas/RequestType"
        confidence:
          type: integer
        status:
          type: string
          enum: [RFQ_CREATED, PENDING_REVIEW, IGNORED]
      required: [id, sender, senderEmail, subject, receivedAt, aiStatus, requestType, confidence, status]
    EmailClassification:
      type: string
      enum: [RFQ, NON_RFQ, SUPPLIER_RESPONSE, CUSTOMER_INQUIRY, INTERNAL, UNKNOWN]
    ProcessingStatus:
      type: string
      enum: [NEW, SYNCED, PROCESSING, PROCESSED, REVIEW_REQUIRED, FAILED, IGNORED]
    AttachmentProcessingStatus:
      type: string
      enum: [PENDING, PROCESSING, EXTRACTED, FAILED, UNSUPPORTED]
    EmailAttachment:
      type: object
      properties:
        id:
          type: integer
        emailId:
          type: integer
        fileName:
          type: string
        contentType:
          type: string
        fileSize:
          type: integer
        processingStatus:
          $ref: "#/components/schemas/AttachmentProcessingStatus"
        extractedText:
          type: ["string", "null"]
      required: [id, emailId, fileName, contentType, fileSize, processingStatus, extractedText]
    AiAnalysis:
      type: object
      properties:
        id:
          type: integer
        emailClassification:
          $ref: "#/components/schemas/EmailClassification"
        requestType:
          $ref: "#/components/schemas/RequestType"
        confidenceScore:
          type: number
        reasoningSummary:
          type: string
        requiresHumanReview:
          type: boolean
        developmentMode:
          type: boolean
        extractedData:
          type: object
          additionalProperties: true
      required: [id, emailClassification, requestType, confidenceScore, reasoningSummary, requiresHumanReview, developmentMode, extractedData]
    ReviewHistoryEntry:
      type: object
      properties:
        id:
          type: integer
        action:
          type: string
          enum: [APPROVED, REJECTED, RECLASSIFIED, EDITED]
        previousClassification:
          $ref: "#/components/schemas/RequestType"
        newClassification:
          $ref: "#/components/schemas/RequestType"
        notes:
          type: string
        createdAt:
          type: string
      required: [id, action, previousClassification, newClassification, notes, createdAt]
    EmailDetail:
      allOf:
        - $ref: "#/components/schemas/Email"
        - type: object
          properties:
            recipient:
              type: string
            cc:
              type: string
            bodyText:
              type: string
            bodyHtml:
              type: string
            processingStatus:
              $ref: "#/components/schemas/ProcessingStatus"
            emailClassification:
              $ref: "#/components/schemas/EmailClassification"
            attachments:
              type: array
              items:
                $ref: "#/components/schemas/EmailAttachment"
            analysis:
              oneOf:
                - $ref: "#/components/schemas/AiAnalysis"
                - type: "null"
            linkedRfq:
              oneOf:
                - $ref: "#/components/schemas/Rfq"
                - type: "null"
          required: [recipient, cc, bodyText, bodyHtml, processingStatus, emailClassification, attachments, analysis, linkedRfq]
...
    EmailProcessResult:
      type: object
      properties:
        email:
          $ref: "#/components/schemas/EmailDetail"
        rfqCreated:
          type: boolean
        reviewRequired:
          type: boolean
        message:
          type: string
      required: [email, rfqCreated, reviewRequired, message]
    AiReviewDetail:
      allOf:
        - $ref: "#/components/schemas/AiReview"
        - type: object
          properties:
            emailSubject:
              type: string
            createdAt:
              type: string
            originalEmail:
              $ref: "#/components/schemas/EmailDetail"
            analysis:
              $ref: "#/components/schemas/AiAnalysis"
            reviewHistory:
              type: array
              items:
                $ref: "#/components/schemas/ReviewHistoryEntry"
          required: [emailSubject, createdAt, originalEmail, analysis, reviewHistory]
    ReclassifyReviewInput:
      type: object
      properties:
        requestType:
          $ref: "#/components/schemas/RequestType"
        notes:
          type: string
      required: [requestType]
    UpdateReviewInput:
      type: object
      properties:
        customer:
          type: string
        partNumber:
          type: string
        description:
          type: string
        quantity:
          type: integer
        condition:
          type: string
        requestType:
          $ref: "#/components/schemas/RequestType"
        priority:
          $ref: "#/components/schemas/Priority"
        notes:
          type: string
      required: []
...
    AiReview:
      type: object
      properties:
        id:
          type: integer
        rfqId:
          type: integer
        rfqNumber:
          type: string
        customer:
          type: string
        classification:
          $ref: "#/components/schemas/RequestType"
        confidence:
          type: integer
        reason:
          type: string
        status:
          type: string
          enum: [NEEDS_REVIEW, APPROVED, CHANGED]
      required: [id, rfqId, rfqNumber, customer, classification, confidence, reason, status]
...