# lib/api-zod/src/generated/types/aiAnalysis.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * M International AI Quote Agent operations API
 * OpenAPI spec version: 0.1.0
 */
import type { AiAnalysisExtractedData } from './aiAnalysisExtractedData';
import type { EmailClassification } from './emailClassification';
import type { RequestType } from './requestType';

export interface AiAnalysis {
  id: number;
  emailClassification: EmailClassification;
  requestType: RequestType;
  confidenceScore: number;
  reasoningSummary: string;
  requiresHumanReview: boolean;
  developmentMode: boolean;
  extractedData: AiAnalysisExtractedData;
}