# lib/api-zod/src/generated/types/processingStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * M International AI Quote Agent operations API
 * OpenAPI spec version: 0.1.0
 */

export type ProcessingStatus = typeof ProcessingStatus[keyof typeof ProcessingStatus];


export const ProcessingStatus = {
  NEW: 'NEW',
  SYNCED: 'SYNCED',
  PROCESSING: 'PROCESSING',
  PROCESSED: 'PROCESSED',
  REVIEW_REQUIRED: 'REVIEW_REQUIRED',
  FAILED: 'FAILED',
  IGNORED: 'IGNORED',
} as const;