# lib/api-zod/src/generated/types/reviewHistoryEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * M International AI Quote Agent operations API
 * OpenAPI spec version: 0.1.0
 */
import type { RequestType } from './requestType';
import type { ReviewHistoryEntryAction } from './reviewHistoryEntryAction';

export interface ReviewHistoryEntry {
  id: number;
  action: ReviewHistoryEntryAction;
  previousClassification: RequestType;
  newClassification: RequestType;
  notes: string;
  createdAt: string;
}