# lib/api-zod/src/generated/types/updateReviewInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * M International AI Quote Agent operations API
 * OpenAPI spec version: 0.1.0
 */
import type { Priority } from './priority';
import type { RequestType } from './requestType';

export interface UpdateReviewInput {
  customer?: string;
  partNumber?: string;
  description?: string;
  quantity?: number;
  condition?: string;
  requestType?: RequestType;
  priority?: Priority;
  notes?: string;
}